refactor: share matrix utils
